ritzy_telegram_store_bot/ritzy_telegram_store_bot/requirements.txt
aiogram>=3.3,<4.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.5.0
python-dotenv>=1.0.1
motor>=3.3.2
httpx>=0.27.0
pytz>=2024.1
